# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Audit Brief API",
    "version": "0.2.1",
...
    "contact": {
      "name": "Audit Brief Platform Team"
    },
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "https://uat.uno.wcgt.in/auditbrief",
      "description": "UAT / production-equivalent (basePath `/auditbrief`)."
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/favorites": {
      "get": {
        "tags": ["Favorites"],
        "operationId": "listFavoriteAuditBriefs",
        "summary": "List favorited audit brief IDs",
        "description": "Authenticated. Returns a **flat** array of audit brief IDs (not\npaginated) because a user's favorites list is expected to remain small\nand is loaded once on the client for UI state hydration.\n",
        "responses": {
          "200": {
            "description": "Array of audit brief IDs favorited by the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "example": ["3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Favorites"],
        "operationId": "toggleFavoriteAuditBrief",
        "summary": "Toggle favorite on an audit brief (idempotent)",
        "description": "Authenticated. Idempotent toggle: creates the favorite (responds `201`,\n`{favorited:true}`) if it does not exist; deletes it (responds `200`,\n`{favorited:false}`) if it does. Races on concurrent toggles are\nswallowed (P2002/P2025) and treated as the desired end-state.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteAuditBriefToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorite removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                },
                "example": {
                  "data": {
                    "favorited": false
                  }
                }
              }
            }
          },
          "201": {
            "description": "Favorite created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                },
                "example": {
                  "data": {
                    "favorited": true
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/learning-graph-favorites": {
      "get": {
        "tags": ["Learning Graph Favorites"],
        "operationId": "listFavoriteLearningGraphs",
        "summary": "List favorited learning graph IDs",
        "description": "Authenticated. Returns a flat array of graph IDs.",
        "responses": {
          "200": {
            "description": "Array of learning graph IDs favorited by the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Learning Graph Favorites"],
        "operationId": "toggleFavoriteLearningGraph",
        "summary": "Toggle favorite on a learning graph (idempotent)",
        "description": "Same idempotent-toggle semantics as `/api/favorites`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteLearningGraphToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorite removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                },
                "example": {
                  "data": {
                    "favorited": false
                  }
                }
              }
            }
          },
          "201": {
            "description": "Favorite created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                },
                "example": {
                  "data": {
                    "favorited": true
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/progress": {
      "get": {
        "tags": ["Progress"],
        "operationId": "listProgress",
        "summary": "List the caller's episode completion records",
        "description": "Authenticated. Ordered by `completedAt` descending. Includes related graph and episode titles.",
        "responses": {
          "200": {
            "description": "Progress records for the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProgressRecord"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Progress"],
        "operationId": "markEpisodeComplete",
        "summary": "Mark an episode as complete (idempotent)",
        "description": "Authenticated. Uses an upsert on the unique constraint `[userId, episodeId]`.\nThe `episodeId` must exist and must belong to `graphId`, otherwise `400`.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Progress record created or updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserProgress"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth v4 JWT session cookie. `httpOnly`, `SameSite=Lax`, and `Secure`\nin production. Over HTTPS the cookie is named\n`__Secure-next-auth.session-token`.\n\nIssued by `POST /api/auth/callback/credentials` (dev) or the Azure AD\nOAuth flow (`GET /api/auth/signin/azure-ad` → provider → `GET /api/auth/callback/azure-ad`).\nContains a signed JWT with `userId`, `role`, `email`, and `jti` (for\nserver-side revocation).\n"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Malformed request, missing required fields, or invalid values.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 400,
              "error_code": "BAD_REQUEST",
              "message": "key parameter required",
              "request_id": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "No session cookie or the session has expired / been revoked.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 401,
              "error_code": "UNAUTHORIZED",
              "message": "Authentication required",
              "request_id": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unhandled server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 500,
              "error_code": "INTERNAL_ERROR",
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["status", "error_code", "message"],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 422
          },
          "error_code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_FAILED",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error classification."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message. Safe to surface to end-users."
          },
          "details": {
            "description": "Free-form diagnostic detail. For `VALIDATION_FAILED` this is a Zod\n`issues` array; for other codes it is commonly absent. May be null.\n"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "Echo of the `x-request-id` header — use for log correlation."
          }
        }
      },
...
      "FavoriteAuditBriefToggleRequest": {
        "type": "object",
        "required": ["auditBriefId"],
        "properties": {
          "auditBriefId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "FavoriteLearningGraphToggleRequest": {
        "type": "object",
        "required": ["learningGraphId"],
        "properties": {
          "learningGraphId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "FavoriteToggleResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["favorited"],
            "properties": {
              "favorited": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "ProgressCreateRequest": {
        "type": "object",
        "required": ["graphId", "episodeId"],
        "properties": {
          "graphId": {
            "type": "string",
            "minLength": 1
          },
          "episodeId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "UserProgress": {
        "type": "object",
        "required": ["id", "userId", "graphId", "episodeId", "completedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "graphId": {
            "type": "string",
            "format": "uuid"
          },
          "episodeId": {
            "type": "string",
            "format": "uuid"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProgressRecord": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserProgress"
          },
          {
            "type": "object",
            "properties": {
              "graph": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              },
              "episode": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}